Allowing capability to be filtered.
Adds a filter, sccss_capability, that lets developers change the required capability to use the plugin. Also fixes a "Cheatin' uh" error if a user with edit_theme_options but not manage_options tries to save the CSS. It does this by returning the filtered capability in the option_page_capability_ hook.